CRYPT-6: Add enemy/player stats and basic attacks
Weapons are used by a player to inflict damage on enemies.

Bleed and toxic are inflicted on attacks and increase and decrease based
on resistances or lack thereof.

Enemies and players share stat calculations. The only exception is that
the DamageMessage for a player attacking an enemy derives the base
damage from the player stats and the weapon itself. For enemies, the
base damage is a direct relationship with the strength stat.